requirements.txt
azure-cosmosdb-table==1.0.6

